rebuild: rework rebuild script, turn it into proper nix package
Previously I thought the only way to use doas in a script was to put it
in ~/.local/bin as a somewhat normal executable script. Now I'm a lil
smarter. The reason why ${pkgs.doas}/bin/doas doesn't work is because it
needs the security.wrapper (NixOS option) features which wrap the
executable with setguid priviliges for certain users and places the
resulting wrapper in /run/wrappers/bin. So you *can* actually use doas
from stuff anywhere in the nix store, you just have to refer to
/run/wrappers/bin/doas directly.

With this knowledge I was able to turn my rebuild script into a standard
shell application in my nixpkgs overlay. The whole point of this
experiement was to be able to use the rebuild script from tmux running
as a systemd user service. Now I can simply include the pkgs.rebuild
bin directory in with the tmux service's PATH.

I also like the fact that I now have no scripts in ~/.local/bin and I
could remove it from my PATH in the future.