overlay: fix self/super usage
Super is for the "old" package set before the overlay is applied and
should be used for callPackage, trivial builders, and referencing
packages you want to replace. Self should be used to refer to the
package set after the overlay is applied. I removed the rec keyword from
the set because it's not necessary.

https://nixos.org/manual/nixpkgs/stable/#sec-overlays-definition